updated error handling

+++ backend/middleware/songrequest.go
@@ -93,64 +93,18 @@ func SongRequest(c *fiber.Ctx) error {
 		})
 	}
 	// Makes the initial DB connection and attempts to create the table
-	db := database.InitializeConnection()
+	db, dbConnErr := database.InitializeConnection()
+	if dbConnErr != nil {
+		return c.JSON(map[string]interface{}{
+			"error": dbConnErr.Error(),
+		})
+	}
 	err := database.CreateTable(query.Channel, db)
 	if err != nil {
 		log.Fatalln(err)
 	}
 	// Attempts to get the latestSongPosition
 	latestSongPos, err := database.GetLatestSongPosition(db, query.Channel)
-
-	// if the table is being created for the first time, the GetLatestSongPosition function can't query through because it thinks that the table was never created so it throws a pq error of undefined_table
-	// so we catch this error and if we do get the "undefined_table" error then create the table "again"(even though it was never created) then insert it
-	if err != nil {
-		if latestSongPos >= 20 {
-			clientData := models.ClientData{
-				Status:  "fail",
-				Message: "The song queue is full!",
-				Data:    nil,
-			}
-			return c.JSON(map[string]interface{}{
-				"error": clientData.Message,
-			})
-		}
-		song := database.ClientSong{
-			User:     query.User,
-			Channel:  query.Channel,
-			Title:    songData.Items[0].Snippet.Title,
-			Artist:   songData.Items[0].Snippet.ChannelTitle,
-			Duration: songDuration.Duration,
-			VideoID:  songData.Items[0].ID.VideoID,
-			Position: latestSongPos + 1,
-		}
-		err := database.CreateTable(query.Channel, db)
-		if err != nil {
-			dataError := database.InsertSong(db, song, query.Channel)
-			if dataError != nil {
-				clientData := models.ClientData{
-					Status:  "fail",
-					Message: dataError.Error(),
-					Data:    nil,
-				}
-				return c.JSON(map[string]interface{}{
-					"error": clientData.Message,
-				})
-			}
-		}
-		
-		clientData := models.ClientData{
-			Status:  "success (table did not exist!)",
-			Message: "inserted into db",
-			Data: []models.Data{
-				{Name: song.Title, Artist: song.Artist, Duration: song.Duration, Position: latestSongPos + 1},
-			},
-		}
-		//insertSong(song)
-		return c.JSON(clientData)
-	}
-
-
-
 
 	if latestSongPos >= 20 {
 		clientData := models.ClientData{
@@ -170,6 +124,15 @@ func SongRequest(c *fiber.Ctx) error {
 		Duration: songDuration.Duration,
 		VideoID:  songData.Items[0].ID.VideoID,
 		Position: latestSongPos + 1,
+	}
+
+	// if the table is being created for the first time, the GetLatestSongPosition function can't query through because it thinks that the table was never created so it throws a pq error of undefined_table
+	// so we catch this error and if we do get the "undefined_table" error then create the table "again"(even though it was never created) then insert it
+	if err != nil {
+		err := database.CreateTable(query.Channel, db)
+		if err != nil {
+			c.Next()
+		}
 	}
 
 	dataError := database.InsertSong(db, song, query.Channel)
@@ -200,7 +163,7 @@ func SongRequest(c *fiber.Ctx) error {
 
 
 
-
+// Middleware function that returns all the songs in that current table.
 func FetchAllSongs(c *fiber.Ctx) error {
 	type Query struct {
 		Channel string `query:"channel"`
@@ -216,7 +179,12 @@ func FetchAllSongs(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
-	db := database.InitializeConnection()
+	db, dbConnErr := database.InitializeConnection()
+	if dbConnErr != nil {
+		return c.JSON(map[string]interface{}{
+			"error": dbConnErr.Error(),
+		})
+	}
 	songs, err := database.GetAllSongRequests(query.Channel, db)
 	if err != nil {
 		return c.JSON(fiber.Map{
@@ -228,6 +196,7 @@ func FetchAllSongs(c *fiber.Ctx) error {
 	})
 }
 
+// Middleware function that deletes the song based of the id passed in the request query.
 func DeleteSong(c *fiber.Ctx) error {
 	type Query struct {
 		Channel string `query:"channel"`
@@ -249,7 +218,12 @@ func DeleteSong(c *fiber.Ctx) error {
 			"error": "missing channel to delete the song from",
 		})
 	}
-	db := database.InitializeConnection()
+	db, dbConnErr := database.InitializeConnection()
+	if dbConnErr != nil {
+		return c.JSON(&fiber.Map{
+			"error": dbConnErr.Error(),
+		})
+	}
 	err := database.DeleteSong(q.Channel, q.Id, db)
 	if err != nil {
 		return c.JSON(&fiber.Map{
